ISC-33863 Winter 2021-22 Plugin updates II, pt.2
