+++ wp-content/plugins/gravityforms/includes/fields/class-gf-field.php
@@ -2534,4 +2534,18 @@ class GF_Field extends stdClass implements ArrayAccess {
 		), esc_html__( 'Quantity:', 'gravityforms' ), $form_id );
 	}
 
+	/**
+	 * Returns an array of key value pairs to be saved to the entry meta after saving/updating the entry.
+	 *
+	 * @since 2.5.16
+	 *
+	 * @param array $form  The form object being saved.
+	 * @param array $entry The entry object being saved
+	 *
+	 * @return array The array that contains the key/value pairs to be stored as extra meta data.
+	 */
+	public function get_extra_entry_metadata( $form, $entry ) {
+		return array();
+	}
+
 }
