+++ wp-content/plugins/gravityforms/includes/legacy/forms_model_legacy.php
@@ -452,7 +452,7 @@ class GF_Forms_Model_Legacy {
 			$field_value = '';
 		}
 
-		self::delete_physical_file( $file_url );
+		self::delete_physical_file( $file_url, $entry_id );
 
 		// update lead field value - simulate form submission
