+++ wp-content/plugins/gravityforms/includes/license/class-gf-license-api-connector.php
@@ -13,7 +13,7 @@ use Gravity_Forms\Gravity_Forms\External_API\GF_API_Response_Factory;
  *
  * Connector providing methods to communicate with the License API.
  *
- * @since 2.5
+ * @since 2.5.11
  *
  * @package Gravity_Forms\Gravity_Forms\License
  */
@@ -43,6 +43,8 @@ class GF_License_API_Connector extends GF_API_Connector {
 	/**
 	 * Check if cache debug is enabled.
 	 *
+	 * @since 2.5.11
+	 *
 	 * @return bool
 	 */
 	public function is_debug() {
@@ -52,7 +54,7 @@ class GF_License_API_Connector extends GF_API_Connector {
 	/**
 	 * If the site was registered with the legacy process.
 	 *
-	 * @since 2.5
+	 * @since 2.5.11
 	 *
 	 * @return bool
 	 */
@@ -62,6 +64,13 @@ class GF_License_API_Connector extends GF_API_Connector {
 
 	}
 
+	/**
+	 * Clear the cache for a given key.
+	 *
+	 * @since 2.5.11
+	 *
+	 * @param string $key
+	 */
 	public function clear_cache_for_key( $key ) {
 		$this->cache->delete( 'rg_gforms_license_info_' . $key );
 	}
@@ -69,8 +78,10 @@ class GF_License_API_Connector extends GF_API_Connector {
 	/**
 	 * Get the license info.
 	 *
-	 * @since 2.5
+	 * @since 2.5.11
 	 *
+	 * @param string $key
+	 * @param bool   $cache
 	 *
 	 * @return GF_API_Response
 	 */
@@ -101,8 +112,10 @@ class GF_License_API_Connector extends GF_API_Connector {
 	/**
 	 * Get the license info.
 	 *
-	 * @since 2.5
+	 * @since 2.5.11
 	 *
+	 * @param string $key
+	 * @param bool   $cache
 	 *
 	 * @return GF_API_Response
 	 */
@@ -133,7 +146,7 @@ class GF_License_API_Connector extends GF_API_Connector {
 	/**
 	 * Check if the saved license key is valid.
 	 *
-	 * @since 2.5
+	 * @since 2.5.11
 	 *
 	 * @return true|WP_Error
 	 */
@@ -147,7 +160,7 @@ class GF_License_API_Connector extends GF_API_Connector {
 	 * Registers a site to the specified key, or if $new_key is blank, unlinks a key from an existing site.
 	 * Requires that the $new_key is saved in options before calling this function
 	 *
-	 * @since 2.5 Implement the license enforcement process.
+	 * @since 2.5.11 Implement the license enforcement process.
 	 *
 	 * @param string $new_key Unhashed Gravity Forms license key.
 	 *
@@ -181,7 +194,9 @@ class GF_License_API_Connector extends GF_API_Connector {
 	/**
 	 * Purge site credentials if the license info contains certain errors.
 	 *
-	 * @since 2.5
+	 * @since 2.5.11
+	 *
+	 * @return void
 	 */
 	public function maybe_purge_site_credentials() {
 
@@ -206,6 +221,8 @@ class GF_License_API_Connector extends GF_API_Connector {
 
 	/**
 	 * Retrieve a list of plugins from the API.
+	 *
+	 * @since 2.5.11
 	 *
 	 * @param bool $cache Whether to respect the cached data.
 	 *
