+++ wp-content/plugins/gravityforms/includes/license/class-gf-license-api-response-factory.php
@@ -9,7 +9,7 @@ use Gravity_Forms\Gravity_Forms\External_API\GF_API_Response_Factory;
  *
  * Concrete response factory used to return a License API Response
  *
- * @since 2.5
+ * @since 2.5.11
  *
  * @package Gravity_Forms\Gravity_Forms\License
  */
@@ -17,12 +17,21 @@ class GF_License_API_Response_Factory implements GF_API_Response_Factory {
 
 	private $transient_strategy;
 
+	/**
+	 * GF_License_API_Response_Factory constructor
+	 *
+	 * @since 2.5.11
+	 *
+	 * @param $transient_strategy
+	 */
 	public function __construct( $transient_strategy ) {
 		$this->transient_strategy = $transient_strategy;
 	}
 
 	/**
 	 * Create a new License API Response from the given data.
+	 *
+	 * @since 2.5.11
 	 *
 	 * @param mixed ...$args
 	 *
